Added support for Ubuntu 22.04

+++ scripts/cleanup-managedimages.ps1
@@ -7,7 +7,8 @@ param(
     [String] [Parameter (Mandatory=$true)] $AgentsResourceGroup,
     [String] [Parameter (Mandatory=$true)] $VmssNameWindows,
     [String] [Parameter (Mandatory=$true)] $VmssNameWindows2022,
-    [String] [Parameter (Mandatory=$true)] $VmssNameUbuntu
+    [String] [Parameter (Mandatory=$true)] $VmssNameUbuntu,
+    [String] [Parameter (Mandatory=$true)] $VmssNameUbuntu2204
 )
 
 az login --service-principal --username $ClientId --password $ClientSecret --tenant $TenantId | Out-Null
@@ -18,9 +19,11 @@ $managedImages = az image list --resource-group $ResourceGroup --subscription $S
 $windowsManagedImage = az vmss show --name $VmssNameWindows --resource-group $AgentsResourceGroup --query 'virtualMachineProfile.storageProfile.imageReference.id' --subscription $SubscriptionId | Out-String | ConvertFrom-Json
 $windowsManagedImage2022 = az vmss show --name $VmssNameWindows2022 --resource-group $AgentsResourceGroup --query 'virtualMachineProfile.storageProfile.imageReference.id' --subscription $SubscriptionId | Out-String | ConvertFrom-Json
 $ubuntuManagedImage = az vmss show --name $VmssNameUbuntu --resource-group $AgentsResourceGroup --query 'virtualMachineProfile.storageProfile.imageReference.id' --subscription $SubscriptionId | Out-String | ConvertFrom-Json
+$ubuntuManagedImage2204 = az vmss show --name $VmssNameUbuntu2204 --resource-group $AgentsResourceGroup --query 'virtualMachineProfile.storageProfile.imageReference.id' --subscription $SubscriptionId | Out-String | ConvertFrom-Json
+
 
 foreach ($managedImage in $managedImages) {
-    if ($managedImage -ne $windowsManagedImage -and $managedImage -ne $ubuntuManagedImage -and $managedImages -ne $windowsManagedImage2022) {
+    if ($managedImage -ne $windowsManagedImage -and $managedImage -ne $ubuntuManagedImage -and $managedImages -ne $windowsManagedImage2022 -and $managedImage -ne $ubuntuManagedImage2204) {
         Write-Host "Found a match, deleting orphaned managed image: $managedImage"
         az image delete --ids $managedImage | Out-Null
     }
